refactor(tests): move async query timeout test to tests/test_async_timeout_recovery.py
